WIP commit.  Merlin 2 integrated.  Currently reworking CLI.
