+++ firebird/__init__.py
@@ -1,89 +1,38 @@
 import logging
+import merlin
 import os
 import socket
 
 # All of these are evaluated at import time!!!  That means the env vars need
 # to be set before the firebird module is imported.  
-HOST = socket.gethostbyname(socket.getfqdn())
-AARDVARK = os.getenv('FIREBIRD_AARDVARK', 'http://localhost:5678')
-AARDVARK_SPECS = os.getenv('FIREBIRD_AARDVARK_SPECS', '/v1/landsat/chip-specs')
-AARDVARK_CHIPS = os.getenv('FIREBIRD_AARDVARK_CHIPS', '/v1/landsat/chips')
-CASSANDRA_CONTACT_POINTS = os.getenv('FIREBIRD_CASSANDRA_CONTACT_POINTS', HOST)
-CASSANDRA_USER = os.getenv('FIREBIRD_CASSANDRA_USER', 'cassandra')
-CASSANDRA_PASS = os.getenv('FIREBIRD_CASSANDRA_PASS', 'cassandra')
-CASSANDRA_KEYSPACE = os.getenv('FIREBIRD_CASSANDRA_KEYSPACE', 'lcmap_changes_local')
-CHIPS_URL = ''.join([AARDVARK, AARDVARK_CHIPS])
-INITIAL_PARTITION_COUNT = int(os.getenv('FIREBIRD_INITIAL_PARTITION_COUNT', 1))
-LOG_LEVEL = os.getenv('FIREBIRD_LOG_LEVEL', 'WARN')
-PRODUCT_PARTITION_COUNT = int(os.getenv('FIREBIRD_PRODUCT_PARTITION_COUNT', 1))
-QA_BIT_PACKED = os.getenv('FIREBIRD_CCD_QA_BITPACKED', 'True')
-SPECS_URL = ''.join([AARDVARK, AARDVARK_SPECS])
-STORAGE_PARTITION_COUNT = int(os.getenv('FIREBIRD_STORAGE_PARTITION_COUNT', 1))
+HOST               = socket.gethostbyname(socket.getfqdn())
+ARD_CHIPMUNK       = os.getenv('ARD_CHIPMUNK', 'http://localhost:5678')
+AUX_CHIPMUNK       = os.getenv('AUX_CHIPMUNK', 'http://localhost:5678')
+CASSANDRA_URLS     = os.getenv('CASSANDRA_URLS', HOST)
+CASSANDRA_USER     = os.getenv('CASSANDRA_USER', 'cassandra')
+CASSANDRA_PASS     = os.getenv('CASSANDRA_PASS', 'cassandra')
+CASSANDRA_KEYSPACE = os.getenv('CASSANDRA_KEYSPACE', 'firebird_local')
+INPUT_PARTITIONS   = int(os.getenv('INPUT_PARTITIONS', 1))
+PRODUCT_PARTITIONS = int(os.getenv('PRODUCT_PARTITIONS', 1))
+LOG_LEVEL          = os.getenv('LOG_LEVEL', 'WARN')
+ARD_CFG            = merlin.cfg.get(profile='chipmunk-ard', env={'CHIPMUNK_URL': ARD_CHIPMUNK}) 
+AUX_CFG            = merlin.cfg.get(profile='chipmunk-aux', env={'CHIPMUNK_URL': AUX_CHIPMUNK}) 
+
 
 # Must obtain a logger from log4j since the jvm is what is actually 
 # doing all the logging under the covers for PySpark.
 # Format and logging configuration is handled through log4j.properties.
 
-def get_logger(sc, name):
+def logger(context, name):
     """Get PySpark configured logger
     
     Args:
-        sc: SparkContext
+        context: SparkContext
         name (str): Name of the logger (category)
 
     Returns:
         Logger instance
     """
 
-    return sc._jvm.org.apache.log4j.LogManager.getLogger(name)
+    return context._jvm.org.apache.log4j.LogManager.getLogger(name)
 
-
-def ccd_params():
-    """Parameters for the ccd algorithm based on the QA_BIT_PACKED environment
-    variable
-
-    Returns:
-        dict: CCD Parameters
-    """
-
-    params = {}
-    if QA_BIT_PACKED is not 'True':
-        params = {'QA_BITPACKED': False,
-                  'QA_FILL': 255,
-                  'QA_CLEAR': 0,
-                  'QA_WATER': 1,
-                  'QA_SHADOW': 2,
-                  'QA_SNOW': 3,
-                  'QA_CLOUD': 4}
-    return params
-
-
-def chip_spec_queries(url):
-    """A map of pyccd spectra to chip-spec queries
-
-    Args:
-        url (str): full url (http://host:port/context) for chip-spec endpoint
-
-    Returns:
-        dict: spectra to chip spec queries
-
-    Example:
-        >>> chip_spec_queries('http://host/v1/landsat/chip-specs')
-        {'reds':     'http://host/v1/landsat/chip-specs?q=tags:red AND sr',
-         'greens':   'http://host/v1/landsat/chip-specs?q=tags:green AND sr'
-         'blues':    'http://host/v1/landsat/chip-specs?q=tags:blue AND sr'
-         'nirs':     'http://host/v1/landsat/chip-specs?q=tags:nir AND sr'
-         'swir1s':   'http://host/v1/landsat/chip-specs?q=tags:swir1 AND sr'
-         'swir2s':   'http://host/v1/landsat/chip-specs?q=tags:swir2 AND sr'
-         'thermals': 'http://host/v1/landsat/chip-specs?q=tags:thermal AND ta'
-         'quality':  'http://host/v1/landsat/chip-specs?q=tags:pixelqa'}
-    """
-
-    return {'reds':     ''.join([url, '?q=tags:red AND sr']),
-            'greens':   ''.join([url, '?q=tags:green AND sr']),
-            'blues':    ''.join([url, '?q=tags:blue AND sr']),
-            'nirs':     ''.join([url, '?q=tags:nir AND sr']),
-            'swir1s':   ''.join([url, '?q=tags:swir1 AND sr']),
-            'swir2s':   ''.join([url, '?q=tags:swir2 AND sr']),
-            'thermals': ''.join([url, '?q=tags:bt AND thermal AND NOT tirs2']),
-            'quality':  ''.join([url, '?q=tags:pixelqa'])}
