+++ firebird/train.py
@@ -1,0 +1,17 @@
+def schema():
+    pass
+
+def inputs():
+    pass
+
+def dataframe():
+    pass
+
+def read():
+    pass
+
+def write():
+    pass
+
+def execute():
+    pass
